flask_chat/requirements.txt
flask
python-dotenv
google-generativeai
langchain
langchain-community
langchain-google-genai
faiss-cpu